+++ loop_step/__init__.py
@@ -8,7 +8,15 @@ A step for loops in the SEAMM flowcharts
 # Bring up the classes so that they appear to be directly in
 # the loop_step package.
 
-from loop_step.loop import Loop  # noqa: F401
+from loop_step.loop import (  # noqa: F401
+    Loop,
+    BreakLoop,
+    break_loop,
+    ContinueLoop,
+    continue_loop,
+    SkipIteration,
+    skip_iteration,
+)
 from loop_step.loop_parameters import LoopParameters  # noqa: F401
 from loop_step.loop_step import LoopStep  # noqa: F401
 from loop_step.tk_loop import TkLoop  # noqa: F401